requirements.txt
# Specific package versions are specified here to allow more consistent caching
# in GitHub Actions. 
#
# I derived this file from the output of the following command and then edited it
# to match the appropriate syntax:
# conda env export > environment.yml
#
# Dependencies from this file can be installed with the following command:
# pip install -r requirements.txt
#
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair>=5.2.0
beautifulsoup4>=4.12.3
black>=24.3.0
colorama
doit>=0.36.0
ipython>=8.22.2
jupyter>=1.0.0
jupyter-book>=1.0.0
jupyterlab>=4.1.5
linearmodels>=5.4
matplotlib>=3.8.3
myst-nb
myst-parser>=2.0.0
notebook>=7.1.2
numpy>=1.26.4
numpydoc>=1.6.0
openpyxl>=3.1.2
pandas>=2.2.1
pandas-datareader>=0.10.0
pandas-market-calendars>=4.4.0
plotly>=5.20.0
plotnine>=0.13.2
polars>=0.20.16
pytest>=8.1.1
python-decouple>=3.8
python-dotenv>=1.0.1
pyxlsb>=1.0.10
requests>=2.31.0
scikit-learn>=1.4.1.post1
scipy>=1.12.0
seaborn>=0.13.2
sphinx-book-theme>=1.1.2
sphinx-autodoc2
statsmodels>=0.14.1
streamlit>=1.32.2
vega_datasets>=0.9.0
wrds>=3.2.0
xbbg>=0.7.7
xlrd>=2.0.1
zstandard>=0.22.0